added product page for individual products, improved admin design and functionality, properly added description to products, added default image for new products

+++ Controllers/admin.js
@@ -20,27 +20,39 @@ module.exports = {
     createProduct: async (req, res) => {
         try {
             await product.create({
+                img: '/imgs/default.jpg',
                 name: 'Name',
                 price: '$0.00',
                 featured: false,
             });
-            res.redirect('/admin');
+            res.status(200).json({ message: "created" })
         } catch (err) {
             console.log(err);
         }
     },
     updateProduct: async (req, res) => {
         try {
-            await product.replaceOne({
-                _id: req.body._id
-            },
-                {
+            let updatedProduct;
+            if (req.file) {
+                updatedProduct = {
                     img: '/uploads/' + req.file.filename,
                     name: req.body.name,
                     price: req.body.price,
+                    description: req.body.description,
                     featured: req.body.featured,
-                });
-            res.redirect('/admin');
+                };
+            } else {
+                updatedProduct = {
+                    name: req.body.name,
+                    price: req.body.price,
+                    description: req.body.description,
+                    featured: req.body.featured,
+                };
+            }
+            await product.updateOne({
+                _id: req.body._id
+            }, updatedProduct);
+            res.status(200).json({ message: "updated" })
         } catch (err) {
             console.log(err);
         }
@@ -50,12 +62,14 @@ module.exports = {
             const productDetails = await product.findOneAndDelete({
                 _id: req.body._id
             });
-            fs.unlink(path.join(__dirname, '..', '/public', productDetails.img), (err) => {
-                if (err) {
-                    throw err;
-                }
-            });
-            res.redirect('/admin');
+            if (productDetails.img && productDetails.img !== '/imgs/default.jpg') {
+                fs.unlink(path.join(__dirname, '..', '/public', productDetails.img), (err) => {
+                    if (err) {
+                        throw err;
+                    }
+                });
+            }
+            res.status(200).json({ message: "deleted" })
         } catch (err) {
             console.log(err);
         }
